tpm2: Cast size_t variable to ssize_t to check for negative values
Cast size_t i to ssize_t i to check for underrun. An underrun could never
occur since it is not possible that all commands are disabled.

+++ src/tpm2/RuntimeCommands.c
@@ -490,7 +490,7 @@ RuntimeCommandsGetArraySize(struct RuntimeCommands *RuntimeCommands)
 
     /* search for byte where command with highest code is enabled */
     for (i = sizeof(RuntimeCommands->enabledCommandsByIdx) - 1;
-         i >= 0;
+         (ssize_t)i >= 0;
          i--) {
         if (RuntimeCommands->enabledCommandsByIdx[i])
             return i + 1;
